Fix the whitespace mapping for typed variables
As we currently don't have any good possibility to store the suffix of a variable name in case of a typed variable (i.e. the whitespace in front of the `:`), we now store it as the variable name's prefix. The prefix will be printed by the containing `J.VariableDeclarations` and we make sure that we don't directly use that for multi-variables. Instead, we then use a wrapping `JS.ScopedVariableDeclarations`.